Refactor DoctorSchedule functionality and DTOs
Significantly refactored the DoctorScheduleController to improve structure and functionality, adding methods for CRUD operations with consistent response handling. Updated PatientsController for correct namespace usage. Introduced CreateDoctorScheduleRequest, CreateDoctorScheduleResponse, DeleteSchedulesResponse, UpdateDoctorScheduleRequest, and UpdateDoctorScheduleResponse classes. Enhanced DoctorScheduleMappingExtensions for better data mapping. Modified GetDoctorSchedule DTO for proper validation. Refactored DoctorScheduleServices for improved error handling and response management. Updated IDoctorSchedule interface to align with new structures. Overall, these changes enhance code organization, readability, and maintainability.
