+++ SmartClinic.API/Controllers/DoctorScheduleController .cs
@@ -1,81 +1,77 @@
-﻿using Microsoft.AspNetCore.Mvc;
-using SmartClinic.Application.Services.Interfaces;
-using SmartClinic.Application.Features.Patients.Query.DTOs.GetDoctorSchedule;
+﻿using SmartClinic.Application.Bases;
 using SmartClinic.Application.Features.DoctorSchedule.Command.CreateDoctorSchedule;
 using SmartClinic.Application.Features.DoctorSchedule.Command.UpdateDoctorSchedule; // إضافة هذا الـ namespace
-using SmartClinic.API.Bases;
-using SmartClinic.Application.Bases;
 using SmartClinic.Application.Features.DoctorSchedule.Command.UpdateDoctorSchedule.SmartClinic.Application.Features.DoctorSchedule.Command.UpdateDoctorSchedule;
+using SmartClinic.Application.Features.DoctorSchedule.Query.DTOs.GetDoctorSchedule;
 
 
-namespace SmartClinic.API.Controllers
+namespace SmartClinic.API.Controllers;
+
+[Route("api/[controller]")]
+[ApiController]
+public class DoctorScheduleController : AppControllerBase
 {
-    [Route("api/[controller]")]
-    [ApiController]
-    public class DoctorScheduleController : AppControllerBase
+    private readonly IDoctorScheduleService _doctorScheduleService;
+
+    public DoctorScheduleController(IDoctorScheduleService doctorScheduleService)
     {
-        private readonly IDoctorScheduleService _doctorScheduleService;
+        _doctorScheduleService = doctorScheduleService;
+    }
 
-        public DoctorScheduleController(IDoctorScheduleService doctorScheduleService)
+    [HttpGet("GetByDoctor/{doctorId}")]
+    [ProducesResponseType<Response<IEnumerable<GetDoctorSchedule>>>(StatusCodes.Status200OK)]
+    [ProducesResponseType(StatusCodes.Status404NotFound)]
+    public async Task<IActionResult> GetScheduleForDoctor(int doctorId)
+    {
+        var response = await _doctorScheduleService.GetSchedulesForDoctorAsync(doctorId);
+
+        if (response == null || !response.Data.Any())
         {
-            _doctorScheduleService = doctorScheduleService;
+            return NotFound(new { message = $"No schedules found for doctor {doctorId}" });
         }
 
-        [HttpGet("GetByDoctor/{doctorId}")]
-        [ProducesResponseType<Response<IEnumerable<GetDoctorSchedule>>>(StatusCodes.Status200OK)]
-        [ProducesResponseType(StatusCodes.Status404NotFound)]
-        public async Task<IActionResult> GetScheduleForDoctor(int doctorId)
+        return NewResult(response);
+    }
+
+    [HttpDelete("{scheduleId}")]
+    [ProducesResponseType<Response<string>>(StatusCodes.Status200OK)]
+    [ProducesResponseType(StatusCodes.Status404NotFound)]
+    public async Task<IActionResult> DeleteScheduleById(int scheduleId)
+    {
+        var response = await _doctorScheduleService.DeleteScheduleByIdAsync(scheduleId);
+
+        if (!response.IsSuccessful)
         {
-            var response = await _doctorScheduleService.GetSchedulesForDoctorAsync(doctorId);
-
-            if (response == null || !response.Data.Any())
-            {
-                return NotFound(new { message = $"No schedules found for doctor {doctorId}" });
-            }
-
-            return NewResult(response);
+            return NotFound(new { message = response.Message });
         }
 
-        [HttpDelete("{scheduleId}")]
-        [ProducesResponseType<Response<string>>(StatusCodes.Status200OK)]
-        [ProducesResponseType(StatusCodes.Status404NotFound)]
-        public async Task<IActionResult> DeleteScheduleById(int scheduleId)
+        return NewResult(new SmartClinic.Application.Bases.Response<string>
         {
-            var response = await _doctorScheduleService.DeleteScheduleByIdAsync(scheduleId);
+            Message = response.Message
+        });
+    }
 
-            if (!response.IsSuccessful)
-            {
-                return NotFound(new { message = response.Message });
-            }
+    [HttpPost]
+    public async Task<IActionResult> CreateDoctorSchedule([FromBody] CreateDoctorScheduleRequest request)
+    {
+        var response = await _doctorScheduleService.CreateAsync(request);
 
-            return NewResult(new SmartClinic.Application.Bases.Response<string>
-            {
-                Message = response.Message
-            });
-        }
-
-        [HttpPost]
-        public async Task<IActionResult> CreateDoctorSchedule([FromBody] CreateDoctorScheduleRequest request)
+        return NewResult(new SmartClinic.Application.Bases.Response<CreateDoctorScheduleResponse>
         {
-            var response = await _doctorScheduleService.CreateAsync(request);
+            Data = response,
+            Message = "Schedule created successfully"
+        });
+    }
 
-            return NewResult(new SmartClinic.Application.Bases.Response<CreateDoctorScheduleResponse>
-            {
-                Data = response,
-                Message = "Schedule created successfully"
-            });
-        }
-  
-        [HttpPut("{scheduleId}")]
-        public async Task<IActionResult> UpdateDoctorSchedule(int scheduleId, [FromBody] UpdateDoctorScheduleRequest request)
+    [HttpPut("{scheduleId}")]
+    public async Task<IActionResult> UpdateDoctorSchedule(int scheduleId, [FromBody] UpdateDoctorScheduleRequest request)
+    {
+        var response = await _doctorScheduleService.UpdateAsync(request);
+
+        return NewResult(new SmartClinic.Application.Bases.Response<UpdateDoctorScheduleResponse>
         {
-            var response = await _doctorScheduleService.UpdateAsync(request);
-
-            return NewResult(new SmartClinic.Application.Bases.Response<UpdateDoctorScheduleResponse>
-            {
-                Data = response,
-                Message = response.Message
-            });
-        }
+            Data = response,
+            Message = response.Message
+        });
     }
 }
