+++ SmartClinic.Application/Features/DoctorsSchedules/Command/CreateDoctorSchedule/CreateDoctorScheduleRequest.cs
