+++ SmartClinic.Application/Features/DoctorsSchedules/Command/DeleteDoctorSchedule/DeleteSchedulesResponse.cs
