+++ SmartClinic.Application/Services/Interfaces/IDoctorSchedule.cs
@@ -1,8 +1,8 @@
 ﻿using SmartClinic.Application.Features.DoctorSchedule.Command.CreateDoctorSchedule;
 using SmartClinic.Application.Features.DoctorSchedule.Command.DeleteDoctorSchedule;
-using SmartClinic.Application.Features.DoctorSchedule.Command.UpdateDoctorSchedule; 
+using SmartClinic.Application.Features.DoctorSchedule.Command.UpdateDoctorSchedule;
 using SmartClinic.Application.Features.DoctorSchedule.Command.UpdateDoctorSchedule.SmartClinic.Application.Features.DoctorSchedule.Command.UpdateDoctorSchedule;
-using SmartClinic.Application.Features.Patients.Query.DTOs.GetDoctorSchedule;
+using SmartClinic.Application.Features.DoctorSchedule.Query.DTOs.GetDoctorSchedule;
 
 namespace SmartClinic.Application.Services.Interfaces
 {
